refactor test handling
I moved TestType to be an enum and created the Test struct which stores
all relevant information about tests. Tests now consist of a TestType, a
name, a function with an argument of anyopaque and args which are those
arguments. This way I only have one struct, but I can still pass around
any runtime or comptime args.

Builtin functions (who are runtime known and take in no arguments) are
kind of jankey. Currently I create a wrapper function that takes in an
anyopaque and assumes that the argument is itself a function which is
then called. This works, and since it's not exposed anywhere it should
be fine, but I should think of a better solution. Or better yet the
lanuguage should make builtin functions comptime known