Factor out printing to terminal
Factoring out printing to the terminal to a seperate struct. This way it
will be easier in the future to change to a buffer I can fully print.

I want to use a buffer so that I don't have to deal with funky ascii
movements anymore. It'd be much easier to fully erase whatever I
printed, update the buffer, and reprint it.

That would also make it much easier to not print intermediate states,
and output everything linearly to output files that do not support ascii