feat: Basic result printer
This is a basic implementation of what I want for the eventual result
printer. On all tests passing or skipped it should currently work
correctly. I'm quite sure it will work on failing regular tests, but it
will fail on failing parameterized tests. In fact, it won't even display
them.

The next steps are to make a special runner for parameterized tests. The
nice thing about this is that this means my main runner doesn't need to
account for running parameterized tests anymore.

The parameterized tests will do some IPC through stdout, which the main
runner will capture, as sort of discussed in #2. I'll make a seperate
issue for it with more detail.