functional IPC
This lays the groundwork for #3.

In this, frankly massive, commit I created the core infrastructure to do
some good IPC beween the test runner and tests themselves.

The infrastructure in use is relatively simple. You have a client and
server, the server takes in an (already spawned) child process to
extract its outputs and inputs. The client extracts it's own stdin and
stdout to communicate over. From there client and server can send one
another messages.

Messages consist out of a 4 byte header, which contains a tag and a byte
len, and afterwards a blob of bytes of length byte len. In the (near)
future I will probably add an structure before the byte len that
indicates some more details about the byte blob

There's currently a proof of concept in test.zig, where a client sends
the server a message, the server replies and then tells the client to
exit.

A grave issue in this commit is the overlap between ClientStreamer and
SeverStreamer, as well as Client and Server in general. I would like to
at least consolidate some of the functions in streamer (like looking for
the delimeter).