ResultPrinter: Completely rework
Internal representation of tests is completely overhauled to properly
store the relation between normal tests and their parameterized tests.

Notably, it is currently not possible to change anything but the most
recent parameterized test for a given parent test. This is because it
was a nightmare to figure out otherwise. I would need to create
something along the lines of an ID for every parameterized test, since
we can't assume that the arg_fmt is unique. I'll maybe have to revisit
this in the future but for now this works for what I'm doing.