Runner: integrate result printer
Finally utilize the result printer in the main runner. With this I was
also able to completely rip out the old test runner, it's not required
anymore. I may add a new TestRunner in the future if it turns out it
makes things more testable, but for now it's just gone.

2 notable issues with this version:
1) Client stdout and stderr is _not_ forwarded to the user
2) It's slow as all hell

I think it's been slow asf since I made all the test run on a different
process, I think even since I just enabled message sending. AKA I'm
pretty sure that sending messages on the client is taking waaay too
long, I'll need to investigate why. A potential solution is to just send
all messages in another thread, so it doesn't block test execution, but
I'll have to see if I can do that/ if that would fix things.